ci: run pr-tile check on `pull_request_target` (#126)
As explain in the README of the action, this check only involves PR
metadata, so it is OK from a security point of view to run it on
`pull_request_target`. This ensures that PR created from a fork
repository received PR comment in case of problems in the PR title.